fix index_list()
the old index_list() was losing reference while it went through the original list change it value for the index.
Now the function lstdup() make a copy of the original list to use as reference.
The old index_list() become the index_manager() and the current index_list() just exist because i went over 25 lines.